requirements.txt
playwright
apscheduler
pyyaml
pymongo
fastapi
uvicorn
